+++ src/main/java/bridge/view/InputView.java
@@ -1,15 +1,22 @@
-package bridge;
+package bridge.view;
+
+import bridge.model.BridgeLength;
+import bridge.util.constant.Messages;
+import camp.nextstep.edu.missionutils.Console;
 
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
  */
 public class InputView {
 
+
+
     /**
      * 다리의 길이를 입력받는다.
      */
-    public int readBridgeSize() {
-        return 0;
+    public static BridgeLength readBridgeSize() {
+        System.out.println(Messages.INPUT_LENGTH_MSG.getMessage());
+        return new BridgeLength(Console.readLine());
     }
 
     /**
